feat: add tags to project being created
I opted for synchronizing the tags selected in between screens using
jotai. The alternative was to somehow wait for the AddTagsScreen to
close to get that information in the previous screen but this seems
cleaner. The only problem I foresee is that we are forced to clean the
selectedTagIds before opening CreateProjectScreen but it's something I
can assume.