feat: add image support in manuals
I tried several things to include imgs in 10tap but nothing worked on a
physical iPhone. The idea was to download images in the app's document
directory, but the webview wouldn't load any picture so I created a new
node (localImage) that renders the base64 of the image. I wonder if this
is going to be a hit on performance once we have enough photos, I hope
not. I reduced the quality of photos from 1 to 0.2 just in case